This application can be used to test new ideas. All the supporting code has been laid out. The only thing that needs to be modified is the **performOperation()** method. This is where the new algorithms/things we want to try go.
This will save us time required to merge new changes checked in by someone while
we have our own changes that we are working on - since they are in different
places. It'll also provide greater safety and comfort to try out new things - since
it does not affect the main program.

NOTE: the new processing algorithms to be tried out go inside "performOperation()";
other helper functions may be added as well. This may be used for personal testing.
Code done here need not be checked in, final check in may still be done in the
main program.